source-code/api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pika
python-dotenv
requests